+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add whitelist for `ssl.SSLContext` (tunnelsociety, #392).
 * Add support for Python 3.14 (even-even).
 * Add more ruff rules (even-even).
+* Add pylint disable format (even-even).
 
 # 2.14 (2024-12-08)