+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-# next (unreleased)
+# 2.15 (2026-03-04)
 
-* Drop support for Python 3.8 (Jendrik Seipp, #398).
 * Handle `while True` loops without `break` statements (kreathon).
 * Add whitelist for `ssl.SSLContext` (tunnelsociety, #392).
-* Add support for Python 3.14 (even-even).
 * Add more ruff rules (even-even).
+* Drop support for Python 3.8 (Jendrik Seipp, #398).
+* Add support for Python 3.14 (even-even).
 
 # 2.14 (2024-12-08)